move promisify to utils
